+++ aura/examples/per-key-effect-2.rs
@@ -9,21 +9,21 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     loop {
         let count = 49;
         for _ in 0..count {
-            *key_colours.key(Key::ROG).0 += 5;
-            *key_colours.key(Key::L).0 += 5;
-            *key_colours.key(Key::I).0 += 5;
-            *key_colours.key(Key::N).0 += 5;
-            *key_colours.key(Key::U).0 += 5;
-            *key_colours.key(Key::X).0 += 5;
+            *key_colours.key(Key::ROG).unwrap().0 += 5;
+            *key_colours.key(Key::L).unwrap().0 += 5;
+            *key_colours.key(Key::I).unwrap().0 += 5;
+            *key_colours.key(Key::N).unwrap().0 += 5;
+            *key_colours.key(Key::U).unwrap().0 += 5;
+            *key_colours.key(Key::X).unwrap().0 += 5;
             writer.write_colour_block(&key_colours)?;
         }
         for _ in 0..count {
-            *key_colours.key(Key::ROG).0 -= 5;
-            *key_colours.key(Key::L).0 -= 5;
-            *key_colours.key(Key::I).0 -= 5;
-            *key_colours.key(Key::N).0 -= 5;
-            *key_colours.key(Key::U).0 -= 5;
-            *key_colours.key(Key::X).0 -= 5;
+            *key_colours.key(Key::ROG).unwrap().0 -= 5;
+            *key_colours.key(Key::L).unwrap().0 -= 5;
+            *key_colours.key(Key::I).unwrap().0 -= 5;
+            *key_colours.key(Key::N).unwrap().0 -= 5;
+            *key_colours.key(Key::U).unwrap().0 -= 5;
+            *key_colours.key(Key::X).unwrap().0 -= 5;
             writer.write_colour_block(&key_colours)?;
         }
     }